Pull request cleanup

+++ tests/epyccel/modules/arrays.py
@@ -546,10 +546,9 @@ def array_int32_in_bool_out_2d_F_complex_3d_expr( x, y, ri ):
 @stack_array('a')
 def array_real_1d_sum_stack_array():
     from numpy import zeros
-    import numpy as np
 
     a = zeros(10)
-    s = np.float64(0.0)
+    s = 0.0
     for i in range(10):
         s += a[i]
     return s
@@ -557,10 +556,9 @@ def array_real_1d_sum_stack_array():
 @stack_array('a')
 def array_real_1d_div_stack_array():
     from numpy import ones
-    import numpy as np
 
     a = ones(10)
-    s = np.float64(0.0)
+    s = 0.0
     for i in range(10):
         s += 1.0 / a[i]
     return s
@@ -569,11 +567,10 @@ def array_real_1d_div_stack_array():
 @stack_array('b')
 def multiple_stack_array_1():
     from numpy import ones, array
-    import numpy as np
 
     a = ones(5)
     b = array([1, 3, 5, 7, 9])
-    s = np.float64(0.0)
+    s = 0.0
     for i in range(5):
         s += a[i] / b[i]
     return s
@@ -582,12 +579,11 @@ def multiple_stack_array_1():
 @stack_array('b', 'c')
 def multiple_stack_array_2():
     from numpy import ones, array
-    import numpy as np
 
     a = ones(5)
     b = array([2, 4, 6, 8, 10])
     c = array([1, 3, 5, 7, 9])
-    s = np.float64(0.0)
+    s = 0.0
     for i in range(5):
         s = s + b[i] - a[i] / c[i]
     return s
@@ -599,10 +595,9 @@ def multiple_stack_array_2():
 @stack_array('a')
 def array_real_2d_sum_stack_array():
     from numpy import zeros
-    import numpy as np
 
     a = zeros((10, 10))
-    s = np.float64(0.)
+    s = 0.
     for i in range(10):
         for j in range(10):
             s += a[i][j]
@@ -611,10 +606,9 @@ def array_real_2d_sum_stack_array():
 @stack_array('a')
 def array_real_2d_div_stack_array():
     from numpy import full
-    import numpy as np
 
     a = full((10, 10), 2)
-    s = np.float64(1.)
+    s = 1.
     for i in range(10):
         for j in range(10):
             s /= a[i][j]
@@ -624,11 +618,10 @@ def array_real_2d_div_stack_array():
 @stack_array('b')
 def multiple_2d_stack_array_1():
     from numpy import ones, array
-    import numpy as np
 
     a = ones((2, 5))
     b = array([[1, 3, 5, 7, 9], [11, 13, 17, 19, 23]])
-    s = np.float64(0.0)
+    s = 0.0
     j = 0
     for i in range(2):
         for j in range(5):
@@ -639,12 +632,11 @@ def multiple_2d_stack_array_1():
 @stack_array('b', 'c')
 def multiple_2d_stack_array_2():
     from numpy import ones, array
-    import numpy as np
 
     a = ones(5)
     b = array([[2, 4, 6, 8, 10], [1, 3, 5, 7, 9]])
     c = array([[1, 3, 5, 7, 9], [2, 4, 6, 8, 10]])
-    s = np.float64(0.0)
+    s = 0.0
     for i in range(2):
         for j in range(5):
             s = s + b[i][j] - a[j] / c[i][j]
