+++ tests/epyccel/recognised_functions/test_numpy_funcs.py
@@ -5272,9 +5272,8 @@ def test_numpy_where_array_like_1d_with_condition(language):
     @types('float64[:]')
     def get_chosen_elements(arr):
         from numpy import where, shape
-        import numpy as np
 
-        a = where(arr > 5, arr, arr * np.float64(2))
+        a = where(arr > 5, arr, arr * 2)
         s = shape(a)
         return len(s), s[0], a[1], a[0]
 
@@ -5482,10 +5481,10 @@ def test_numpy_linspace_scalar(language):
     @types('float64', 'int', 'int')
     def get_linspace(start, steps, num):
         from numpy import linspace
-        import numpy as np
+
         stop = start + steps
         b = linspace(start, stop, num)
-        x = np.float64(0.0)
+        x = 0.0
         for bi in b:
             x += bi
         return x
